Negative tests for registration created All tests have been moved to different classes
